Bound the live Raft channels and use fixed-width public integers
Clear every unbounded_channel and usize_in_public_api finding. The replica inbox is bounded by the admitted time profile's scheduler queue depth and timer tasks await delivery (delayed, never dropped); the supervision control port uses try_send and returns an explicit full-queue denial without enqueueing. Public signatures take and return u64/u32 with checked crate::bounded conversions and labelled overflow errors, no as casts. Pinned Octet: unbounded_channel 9->0, usize_in_public_api 32->0 (workspace), 16->0 (lib), no other family changed; fmt, clippy -D warnings, and 2091 workspace tests pass (fabricboundarycompat 5/5); harness suites and fabric-simulation run/shrink are byte-identical to the base.

+++ src/capability/tokens.rs
@@ -10,7 +10,7 @@ const CHECK_STATUS_PASS: &str = "pass";
 const CHECK_STATUS_FAIL: &str = "fail";
 const DECISION_PASS: &str = "pass";
 const DECISION_DENY: &str = "deny";
-const MAX_UCAN_RECEIPT_REFS: usize = 1024;
+const MAX_UCAN_RECEIPT_REFS: u64 = 1024;
 
 type UcanCheckEvaluator = fn(&UcanVerificationChecks) -> bool;
 type UcanCheckSpec = (&'static str, UcanCheckEvaluator);
@@ -485,7 +485,7 @@ fn ucan_verification_diagnostics(input: &UcanVerificationInput) -> Vec<String> {
     diagnostics
 }
 
-const UCAN_VERIFICATION_RECEIPT_ARITY: usize = 19;
+const UCAN_VERIFICATION_RECEIPT_ARITY: u64 = 19;
 const UCAN_VERIFICATION_SCHEMA_INDEX: usize = 0;
 const UCAN_VERIFICATION_DECISION_INDEX: usize = 1;
 const UCAN_VERIFICATION_DIAGNOSTICS_INDEX: usize = 2;
